refactor: sorting files by folders
