+++ src/modules/albums/services/album-api.ts
@@ -1,5 +1,5 @@
 import { RESTDataSource, RequestOptions } from 'apollo-datasource-rest';
-import { IAlbum } from './album-type';
+import { IAlbum } from '../album-type';
 
 export class AlbumAPI extends RESTDataSource {
   constructor() {
@@ -7,7 +7,7 @@ export class AlbumAPI extends RESTDataSource {
     this.baseURL = process.env.ALBUMS_URL;
   }
   willSendRequest(request: RequestOptions) {
-    request.headers.set('Authorization', `Bearer ${this.context.token}`);
+    request.headers.set('Authorization', this.context.token);
   }
   async getAlbum(albumId: string) {
     return this.get(`/${albumId}`);
