+++ src/modules/artists/artist-type.ts
