+++ src/modules/bands/band-type.ts
