+++ src/modules/favourites/favourites-type.ts
